change formating of SQL statements to fix TYPO3 database compare
TYPO3 offers in the install tool an action "Database analyzer" which is
used regularly by admins to check if SQL-specification of extensions fit
the current database.

In case of DPF there are multiple rows which should be changed but
executing the queries in the analyzer changes nothing at all.

The reason is simple. TYPO3 requires some special formating of the SQL
statements:
* data types have to be written in lowercase
* no spaces are allowed in KEY() statements

This patch change the formating of ext_tables.sql. The "Database
analyzer" offers no necessary changes now.