Configurable allowed API actions
The plugins API controller checks for a setting called
`allowedActions`. This setting should be an array of actions that
should be allowed for the given request.

The setting can be configured using TYPOScript conditions in the page setup.
A default setup is provided allowing the actions `mets`, `preview` and
`dataCite` only to development IPs or requests from localhost. The
`attachment` action is always allowed.

The controller will respond with 403 Forbidden to any request to actions
not listed in `plugin.tx_dpf.settings.allowedActions`.