Overwrite backend layouts as of TYPO3 7.6
By the configured typoscript condition the path to backend layouts gets
overwritten for TYPO3 >=7.6.

Using the plural constants for templateRootPath(s), partialRootPath(s)
and layoutRootPath(s) makes it possible to have a fallback. All
templates which does not exist in highest numbered folder are search in
the next lower.

-->

module.tx_dpf.view.layoutRootPaths.100 overwrites
module.tx_dpf.view.layoutRootPaths.10

But e.g. AdminBE.html which is not present in layoutRootPaths.100 will
be searched in layoutRootPaths.10.

This feature exists as of TYPO3 6.2.