Use extension settings even if no client settings are available.
Setting for e.g. "elasticSearchHost" may be done in client configuration
record or in system-wide extension configuration. If a client setting is
present it will be used over the system-wide setting.

If no client configuration is present, the system-wide configuration
should be used. This was not the case in current code.