Replace custom Docker environment with DDEV setup (#174)
DDEV-typo3 is setup around the idea of a full TYPO3 project where
everything is controlled by a central `.ddev/config.yaml` and
composer.json. So the code for the actual Kitodo.Publication extension
has to go into its own `src` directory.

The `src` directory is configured to be a repository root path for
kitodo-publication, so that it is symlinked into typo3conf/ext on
installation.

The commit includes an initial database dump and a custom `ddev test`
command for running unit tests.