Fix form crash on save with invalid values
When hitting the save button Javascript codes validates the form fields
and blocks the save button to dismiss the save action event. When form
validation errors are detected the form input tabs (pages) are marked
with an error icon (flame).

The selection expression was invalid causing the Javascript validation
function to crash. Subsequently the save action is not blocked and
invalid tabs are not marked.

Resolves https://jira.slub-dresden.de/browse/CMR-995
Resolves https://jira.slub-dresden.de/browse/CMR-954