Allow TYPO3 composer plugins
As of Composer 2.2. 0, the allow-plugins option adds a layer
of security allowing you to restrict which Composer plugins
are able to execute code during a Composer run. When a new
plugin is first activated, which is not yet listed in the
config option, Composer will print a warning.

See https://getcomposer.org/doc/06-config.md#allow-plugins