Fix form rendering error on missing item IDs
The form mechanism applies IDs to every form group and field in order to
properly read, update and compare forms that are based on the same
configuration.

However, if input XML does not come with such form field IDs, the
DocumentMapper class has to set new IDs on the fly.

This commit fixes a problem with setting up form IDs if they do not
exist already.

Resolves https://jira.slub-dresden.de/browse/CMR-1255